refactor: move code in generate-registry to asset package
